requirements.txt
groq
phidata
OpenAI
yfinance
duckduckgo-search